AG-12847 - column resize module
