+++ community-modules/core/src/columns/columnAutosizeService.ts
@@ -5,13 +5,19 @@ import type { CtrlsService } from '../ctrlsService';
 import type { AgColumn } from '../entities/agColumn';
 import type { AgColumnGroup } from '../entities/agColumnGroup';
 import type { ColumnEventType } from '../events';
+import type { ScrollVisibleService } from '../gridBodyComp/scrollVisibleService';
 import type { HeaderGroupCellCtrl } from '../headerRendering/cells/columnGroup/headerGroupCellCtrl';
+import type { IColumnLimit, ISizeColumnsToFitParams } from '../interfaces/autoSize';
 import type { IRenderStatusService } from '../interfaces/renderStatusService';
 import type { AnimationFrameService } from '../misc/animationFrameService';
 import type { AutoWidthCalculator } from '../rendering/autoWidthCalculator';
+import { _removeFromArray } from '../utils/array';
+import { _getInnerWidth } from '../utils/dom';
+import { _warnOnce } from '../utils/function';
 import { TouchListener } from '../widgets/touchListener';
 import type { ColumnEventDispatcher } from './columnEventDispatcher';
 import type { ColKey, ColumnModel, Maybe } from './columnModel';
+import { getWidthOfColsInList } from './columnUtils';
 import type { VisibleColsService } from './visibleColsService';
 
 export class ColumnAutosizeService extends BeanStub implements NamedBean {
@@ -24,6 +30,8 @@ export class ColumnAutosizeService extends BeanStub implements NamedBean {
     private eventDispatcher: ColumnEventDispatcher;
     private ctrlsService: CtrlsService;
     private renderStatusService?: IRenderStatusService;
+    private scrollVisibleService: ScrollVisibleService;
+
     private timesDelayed = 0;
 
     public wireBeans(beans: BeanCollection): void {
@@ -33,7 +41,8 @@ export class ColumnAutosizeService extends BeanStub implements NamedBean {
         this.autoWidthCalculator = beans.autoWidthCalculator;
         this.eventDispatcher = beans.columnEventDispatcher;
         this.ctrlsService = beans.ctrlsService;
-        this.renderStatusService = beans.renderStatusService as IRenderStatusService;
+        this.renderStatusService = beans.renderStatusService;
+        this.scrollVisibleService = beans.scrollVisibleService;
     }
 
     public postConstruct(): void {
@@ -231,6 +240,206 @@ export class ColumnAutosizeService extends BeanStub implements NamedBean {
         element.addEventListener('dblclick', listener);
 
         return () => element.removeEventListener('dblclick', listener);
+    }
+
+    // method will call itself if no available width. this covers if the grid
+    // isn't visible, but is just about to be visible.
+    public sizeColumnsToFitGridBody(params?: ISizeColumnsToFitParams, nextTimeout?: number): void {
+        const gridBodyCtrl = this.ctrlsService.getGridBodyCtrl();
+        const removeScrollWidth = gridBodyCtrl.isVerticalScrollShowing();
+        const scrollWidthToRemove = removeScrollWidth ? this.scrollVisibleService.getScrollbarWidth() : 0;
+        // bodyViewportWidth should be calculated from eGridBody, not eBodyViewport
+        // because we change the width of the bodyViewport to hide the real browser scrollbar
+        const bodyViewportWidth = _getInnerWidth(gridBodyCtrl.getGridBodyElement());
+        const availableWidth = bodyViewportWidth - scrollWidthToRemove;
+
+        if (availableWidth > 0) {
+            this.sizeColumnsToFit(availableWidth, 'sizeColumnsToFit', false, params);
+            return;
+        }
+
+        if (nextTimeout === undefined) {
+            window.setTimeout(() => {
+                this.sizeColumnsToFitGridBody(params, 100);
+            }, 0);
+        } else if (nextTimeout === 100) {
+            window.setTimeout(() => {
+                this.sizeColumnsToFitGridBody(params, 500);
+            }, 100);
+        } else if (nextTimeout === 500) {
+            window.setTimeout(() => {
+                this.sizeColumnsToFitGridBody(params, -1);
+            }, 500);
+        } else {
+            _warnOnce(
+                'tried to call sizeColumnsToFit() but the grid is coming back with ' +
+                    'zero width, maybe the grid is not visible yet on the screen?'
+            );
+        }
+    }
+
+    // called from api
+    public sizeColumnsToFit(
+        gridWidth: any,
+        source: ColumnEventType = 'sizeColumnsToFit',
+        silent?: boolean,
+        params?: ISizeColumnsToFitParams
+    ): void {
+        if (this.columnModel.isShouldQueueResizeOperations()) {
+            this.columnModel.pushResizeOperation(() => this.sizeColumnsToFit(gridWidth, source, silent, params));
+            return;
+        }
+
+        const limitsMap: { [colId: string]: Omit<IColumnLimit, 'key'> } = {};
+        if (params) {
+            params?.columnLimits?.forEach(({ key, ...dimensions }) => {
+                limitsMap[typeof key === 'string' ? key : key.getColId()] = dimensions;
+            });
+        }
+
+        // avoid divide by zero
+        const allDisplayedColumns = this.visibleColsService.allCols;
+
+        const doColumnsAlreadyFit = gridWidth === getWidthOfColsInList(allDisplayedColumns);
+        if (gridWidth <= 0 || !allDisplayedColumns.length || doColumnsAlreadyFit) {
+            return;
+        }
+
+        const colsToSpread: AgColumn[] = [];
+        const colsToNotSpread: AgColumn[] = [];
+
+        allDisplayedColumns.forEach((column) => {
+            if (column.getColDef().suppressSizeToFit === true) {
+                colsToNotSpread.push(column);
+            } else {
+                colsToSpread.push(column);
+            }
+        });
+
+        // make a copy of the cols that are going to be resized
+        const colsToDispatchEventFor = colsToSpread.slice(0);
+        let finishedResizing = false;
+
+        const moveToNotSpread = (column: AgColumn) => {
+            _removeFromArray(colsToSpread, column);
+            colsToNotSpread.push(column);
+        };
+
+        // resetting cols to their original width makes the sizeColumnsToFit more deterministic,
+        // rather than depending on the current size of the columns. most users call sizeColumnsToFit
+        // immediately after grid is created, so will make no difference. however if application is calling
+        // sizeColumnsToFit repeatedly (eg after column group is opened / closed repeatedly) we don't want
+        // the columns to start shrinking / growing over time.
+        //
+        // NOTE: the process below will assign values to `this.actualWidth` of each column without firing events
+        // for this reason we need to manually dispatch resize events after the resize has been done for each column.
+        colsToSpread.forEach((column) => {
+            column.resetActualWidth(source);
+
+            const widthOverride = limitsMap?.[column.getId()];
+            const minOverride = widthOverride?.minWidth ?? params?.defaultMinWidth;
+            const maxOverride = widthOverride?.maxWidth ?? params?.defaultMaxWidth;
+
+            const colWidth = column.getActualWidth();
+            if (typeof minOverride === 'number' && colWidth < minOverride) {
+                column.setActualWidth(minOverride, source, true);
+            } else if (typeof maxOverride === 'number' && colWidth > maxOverride) {
+                column.setActualWidth(maxOverride, source, true);
+            }
+        });
+
+        while (!finishedResizing) {
+            finishedResizing = true;
+            const availablePixels = gridWidth - getWidthOfColsInList(colsToNotSpread);
+            if (availablePixels <= 0) {
+                // no width, set everything to minimum
+                colsToSpread.forEach((column) => {
+                    const widthOverride = limitsMap?.[column.getId()]?.minWidth ?? params?.defaultMinWidth;
+                    if (typeof widthOverride === 'number') {
+                        column.setActualWidth(widthOverride, source, true);
+                        return;
+                    }
+                    column.setMinimum(source);
+                });
+            } else {
+                const scale = availablePixels / getWidthOfColsInList(colsToSpread);
+                // we set the pixels for the last col based on what's left, as otherwise
+                // we could be a pixel or two short or extra because of rounding errors.
+                let pixelsForLastCol = availablePixels;
+                // backwards through loop, as we are removing items as we go
+                for (let i = colsToSpread.length - 1; i >= 0; i--) {
+                    const column = colsToSpread[i];
+
+                    const widthOverride = limitsMap?.[column.getId()];
+                    const minOverride = widthOverride?.minWidth ?? params?.defaultMinWidth;
+                    const maxOverride = widthOverride?.maxWidth ?? params?.defaultMaxWidth;
+                    const colMinWidth = column.getMinWidth();
+                    const colMaxWidth = column.getMaxWidth();
+                    const minWidth =
+                        typeof minOverride === 'number' && minOverride > colMinWidth ? minOverride : colMinWidth;
+                    const maxWidth =
+                        typeof maxOverride === 'number' && maxOverride < colMaxWidth ? maxOverride : colMaxWidth;
+                    let newWidth = Math.round(column.getActualWidth() * scale);
+
+                    if (newWidth < minWidth) {
+                        newWidth = minWidth;
+                        moveToNotSpread(column);
+                        finishedResizing = false;
+                    } else if (newWidth > maxWidth) {
+                        newWidth = maxWidth;
+                        moveToNotSpread(column);
+                        finishedResizing = false;
+                    } else if (i === 0) {
+                        // if this is the last column
+                        newWidth = pixelsForLastCol;
+                    }
+
+                    column.setActualWidth(newWidth, source, true);
+                    pixelsForLastCol -= newWidth;
+                }
+            }
+        }
+
+        // see notes above
+        colsToDispatchEventFor.forEach((col) => {
+            col.fireColumnWidthChangedEvent(source);
+        });
+
+        this.visibleColsService.setLeftValues(source);
+        this.visibleColsService.updateBodyWidths();
+
+        if (silent) {
+            return;
+        }
+
+        this.eventDispatcher.columnResized(colsToDispatchEventFor, true, source);
+    }
+
+    public applyAutosizeStrategy(): void {
+        const autoSizeStrategy = this.gos.get('autoSizeStrategy');
+        if (!autoSizeStrategy) {
+            return;
+        }
+
+        const { type } = autoSizeStrategy;
+        // ensure things like aligned grids have linked first
+        setTimeout(() => {
+            if (type === 'fitGridWidth') {
+                const { columnLimits: propColumnLimits, defaultMinWidth, defaultMaxWidth } = autoSizeStrategy;
+                const columnLimits = propColumnLimits?.map(({ colId: key, minWidth, maxWidth }) => ({
+                    key,
+                    minWidth,
+                    maxWidth,
+                }));
+                this.sizeColumnsToFitGridBody({
+                    defaultMinWidth,
+                    defaultMaxWidth,
+                    columnLimits,
+                });
+            } else if (type === 'fitProvidedWidth') {
+                this.sizeColumnsToFit(autoSizeStrategy.width, 'sizeColumnsToFit');
+            }
+        });
     }
 
     // returns the width we can set to this col, taking into consideration min and max widths
