+++ community-modules/core/src/context/gridBeanComparator.ts
@@ -90,7 +90,8 @@ const orderedCoreBeans: BeanName[] = [
     'columnAutosizeService',
     'columnGetStateService',
     'columnGroupStateService',
-    'columnSizeService',
+    'columnFlexService',
+    'columnResizeService',
     'funcColsService',
     'columnNameService',
     'columnViewportService',
